Types

Index

EnergyModelsCO2.AccumulatingStrategic

EnergyModelsCO2.CaptureNone

Parameter types

EnergyModelsCO2.AccumulatingStrategic — Type

AccumulatingStrategic <: EMB.Accumulating

StorageBehavior which accumulates all inflow witin a strategic period and transfers the level to the next strategic period. This approach is used for CO2Storage nodes.

Legacy type constructor

EnergyModelsCO2.CaptureNone — Function

CaptureNone(emissions::Dict{ResourceEmit,T}, co2_capture::Float64)

Legacy constructor for a CaptureNone. This type was renamed to CaptureFlueGas while everything else remains the same.