Methods - Accessing fields

Index

EnergyModelsCO2.co2_proxy

Node

EnergyModelsCO2.co2_proxy — Function

co2_proxy(n::EMB.Node)

Extract the instance of the CO₂ proxy. This function is available for all node types but will provide an error if the node type does not support CO₂ capture retrofit.